WifiService: Use WifiInfo for masking
Migrate to using the new |parcelSensitiveFields| param for masking
location data in WifiInfo. This will ensure that we have a consistent
mechanism for masking in the legacy and new API's for retrieving
WifiInfo.

Note: This change is not dependent on the device SDK (even though the
change is in an API class - WifiInfo).

Bug: 162602799
Test: atest com.android.server.wifi
Change-Id: Ic68546e2a409aeb1f7b75e77d70e7f675e14f3ce
Merged-In: Ic68546e2a409aeb1f7b75e77d70e7f675e14f3ce