Optimize the process of toggle off WiFi
When WiFi is toggled off and on again quickly, it may fail to toggle on.
The current progress of toggle off is to send the WiFi disabled
broadcast first,and then continue to toggle off WiFi. If at this time,
the user toggle on the WiFi, the message may be abort because the
ClientModeManager has quit the StateMachine. The normal process
should be to toggle off WiFi first, and then send the broadcast.

Signed-off-by: zhujiatai <zhujiatai@xiaomi.com>
Change-Id: I84b1c28b9f4526af08c60dec830a47fd24c7b1d4