Tolerate a failure to start the WPA supplicant.
Currently with emulated devices such as FVP that lack wifi but declare
support for it in order to prevent Context.getSystemService("wifi")
from throwing an exception, we attempt to launch the wpa_supplicant
service if we see a wifi device. This can fail if the service is
not available. Currently we handle the failure by bringing down the
entire system_server, which prevents the system from coming up at
all. In  this case we should tolerate the failure and let the rest
of the system come up, so return false from startDaemon(), similar
to how we already handle the V1.1 case.

Bug: 142352330
Change-Id: I22ab306edd1c49792e4cb940a18bcc407519db1a