Move test into nested class
When using @RunWith(Enclosed.class), tests in the outer
class won't be run. There was even a warning comment
about this on this test class, but it was still missed.

Errorprone is being updated to 3.8.1, and adds a check for
this.

Bug: 190944875
Test: Presubmits
Change-Id: I4e53fe9a157c2daba32ca4c77895b37c6723a71d
Merged-In: Ie08b155e2e6061f91575b64126e643e732b62359