Fix the occasional occurrence of invalid WiFi status bar signal
Bug:225996194

Occasionally the RSSI obtained once is invalid (-127). The OnCap update
is triggered and the WiFi signal update in the status bar is invalid.
but the mLastSignalLevel is not updated.When the RSSI is subsequently
obtained correctly. since mLastSignalLevel is kept the same as newSignalLevel
OnCap will not be triggered and the WiFi signal in the status bar remains invalid.

Change-Id: I336d6a70fe07078e1d945dabfdb82355992b1413
Signed-off-by: dongziqi <dongziqi1@xiaomi.corp-partner.google.com>