Ignore other requests while ClientModeManager is starting
Bug: 239887781

When CMD_WIFI_TOGGLED is received, the ClientModeManager with the role
ROLE_CLIENT_PRIMARY is started, but the role is not set successfully
immediately. If CMD_SCAN_ALWAYS_MODE_CHANGED is received at this time,
the ClientModeManager with the role ROLE_CLIENT_PRIMARY will be
started again, resulting in an incorrect WiFi state.

Change-Id: Id967b35e20c2012aa8b6dc3c81ea421951b46816
Signed-off-by: DingFei Song <v-dingfeisong@xiaomi.corp-partner.google.com>