Revert "Merge TP1A.221105.002 to aosp-master - DO NOT MERGE"
Revert "Merge TP1A.221105.002 to aosp-master - DO NOT MERGE"

This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f6eb4e9-d61f-4a92-8c74-4e9098a71713). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I97a17772c49459881fd1c80e0b7b669f098a3d5e