Fix the memory leak caused by softapmanager's failure to unregister after registration with ClientModeImplMonitor
Reason: Softapmanager failed to unregister after being registered
in ClientModeImplMonitor, the discarded softapmanager cannot be
recycled after multiple hotspots are switched on and off, which
will lead to memory leaks, the onL2connected interface is executed
continuously after the wifi connection. If other calls are added to
the interface, Repetitive execution of the calling logic may cause
system exceptions.

Solution: refer to WifiNetworkFactory to unregister before closing
the hotspot

Bug:260667331

Change-Id: Ia6fdbdde0687c11c1466d727d25f2187a78154cb
Signed-off-by: daichen3 <daichen3@xiaomi.com>