wifi: Disconnect on IP_REACHABILITY_LOST for a specific period
IP_REACHABILITY_LOST shall be triggered on a NUD_FAIL indication.
The current logic to check for the IP Reachability is to wait for
the response for the configured number of ARP requests to the
gateway before triggering a NUD_FAILURE. This feature when enabled
by default is resulting into some false positives (due to the ARP
responses getting delayed from the AP), leading to a disconnection
 / reconnection to the network.

Considering the pros / cons of this feature, this commit customizes
this behavior to be active only for the first 10 seconds after the
connect/roam to the AP. This should help to identify the roaming
across multiple subnets.The new architectures have offloaded this
IPReachabilityMonitor to the host driver which have more control of
the Wi-Fi statistics before they trigger the disconnect. This should
address any issue due to the data stalls.

All said,this enhancement of "Disconnect on IP_REACHABILITY_LOST for a
specific period" is enabled by default and controlled by the overlay
"config_disconnectOnlyOnInitialIpReachability" .
Set this to false if this enhancement is not required in device overlay.

Bug: 262321154
Test: atest -c FrameworksWifiTests
Change-Id: I518f8e6372fb0edcf86ad006cf101e58b471419c
Signed-off-by: xuelin <xuelin@xiaomi.corp-partner.google.com>