Wifi: Ignore invalid SSIDs and don't crash system.
In case SSID is invalid the constructor of WifiSsid throws
an IllegalArgumentException with text, e.g.:
Max SSID length is 32 bytes, but received 64 bytes!

This is currently handled as fatal error which leads
to an complete restart of the system.

Change the behavior to log this error and continue
with the next NativeScanResult entry of the scanned
result list.

Change-Id: I21a6821c8e5ebff61733f11f2b25d22d3149f9e0