Repair hybrid encryption AP first close automatic connection disconnection
When connecting to a WPA2/WPA3 AP for the first time, set IsAddedByAutoUpgrade of WPA3 to true.
Set IsAddedByAutoUpgrade of WPA3 to false after turning off automatic connection for the first time and then saving.
Causes credentials to change, requiring a reconnect. This value should not affect credentials and is recommended to be ignored.

Change-Id: I99e03963d92ad430bc40a7d0a724d6686d23ef6b
Signed-off-by: Yu <zhangyu34@xiaomi.com>