Expose framework-wifi.impl to //vendor
The private apis of framwork-wifi are currently being accessed
by some vendor code, via hacks that use makefiles and the flattened
version of the wifi apex. We want to remove flattened apexes and
convert makefiles to blueprint files, so we'll convert the vendor
code to access the wifi private apis via framework-wifi.impl instead.
This requires exposing the impl library to //vendor.

Bug: 254205429
Bug: 200577024
Test: Presubmits
Change-Id: I975deeda1600e0d8443a596c4ec349b0f091579e