Add onInterfaceClassActivityChangedParcel to INetdUnsolicitedEventListener
Currently, onInterfaceClassActivityChanged is used to track default
interface activity.
On U+ devices, netd calls onInterfaceClassActivityChangedParcel
instead and interface index is newly added to this callback argument.
This change is a preparation to track all the interface activity.

onInterfaceClassActivityChanged is not used in wifi, so this change does
not have any impact

Test: netd_integration_test
Bug: 270298713
Change-Id: I31871218c15a8382362df5d86eed52da7a841ae4