[Bugfix][wifi] fix wrong passwd could connect wpa3-psk successfully issue.
MIUIROM-1518045

Preshare psk will be mark *, so just modify passwd, it will not trigger
to delete pmk cache in WifiServce process. Then when start to connect
the AP, WifiServce will set pmk cache info to wpa supplicant, it will
connect the AP successfully, because of that pmk cache is right.

Signed-off-by: fengchao <fengchao@xiaomi.com>
Change-Id: I816a0bf4e4148a389a9432640ba6ce8aa6c784b3