Add Tethering APEX as a dependency of the Wifi bootclasspath-fragment
Some classes in the Wifi APEX (e.g. WifiInfo, WifiNetworkSpecifier)
depends on the API provided by the Tethering APEX (e.g.
NetworkSpecifier), but Tethering APEX is currently not listed as a
dependency of the Wifi bootclasspath-fragment, leading to the hiddenapi
flags being not properly propagated to the child classes and causing
difference between the generated hiddenapi entries between the from-text
stub build and from-source stub build.

This change fixes such problem by adding the Tethering apex as the
dependency of the Wifi bootclasspath fragment.

Test: m --build-from-text-stub && manually inspect the diff in hiddenapi-flags.csv
Bug: 308501909
Change-Id: I5b46ffdb5644e6f4bb0dc6df951de1f397cfeab3