Use appropriate API surface in framework-wifi dependencies
Previously, `framework-wifi` had the following:
```
stub_only_libs: ["framework-connectivity.stubs.module_lib"],
```

That meant that all of the `framework-wifi` API surfaces depended on
the module-lib API surface from `framework-connectivity` which meant
that references to `module-lib` classes could leak into the `public`
and `system` API surfaces. This change fixes that by making sure that
each of the `framework-`wifi` API surfaces depends on the corresponding
API surface from `framework-connectivity`.

Usually that would be done simply by adding:
```
stub_only_libs: ["framework-connectivity"],
```

But that does not work as it causes a dependency cycle because the
underlying `framework-connectivity` `java_library` depends upon
`framework-wifi.stubs.module_lib` which depends upon
`framework-connectivity`.

The above change has the intentional side effect of causing the
`WifiInfo.getApplicableRedactions()` to be added to the `current.txt`.
Previously, it was not included in the `current.txt` file because it
overrides `TransportInfo.getApplicableRedactions()` which is present
in the `framework-connectivity`'s `module-lib` API. However, that
method is NOT in the `framework-connectivity`'s `public` API.

It was also necessary to remove
`framework-connectivity.stubs.module_lib` from the `libs` property in
`framework-wifi-defaults` as that would end up being passed to
`framework-wifi`'s `libs` property which would have a similar effect
on metalava. Instead it was passed to `impl_only_libs` in
`framework-wifi` and `libs` in `framework-wifi-pre-jarjar`.

Bug: 308983623
Test: m checkapi
Change-Id: I35435378f06591be754d8d05f48c9b80f8efaf65