Revert "Inline literal string value of Flags in API signature files"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3972df80-a1a8-4fa8-bfc4-a7254147ae4c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1d3f22b947df396612a2e9907b35ea6c6e78da91