Revert "Inline literal string value of Flags in API signature files"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1a719dff-dd59-48b0-b1a1-43c05e052ce7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib6b0511db0c2bae9f410330d43888af60ad717f7