Revert "Mock Wifi:Change isPnoSupported in WifiServiceImp to ref..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5bfdcf2d-8e1b-45fb-a717-c7045a98b433). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0c3024180599956ca3e089ef4551f1f8835db898