Revert "Use certificate build target"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/90ce5c2f-be6c-4d97-bae1-3dee6fc43912). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9f5139a51db0968a1e3cc36d683f58742f705f79