Fix：Very fast switching of GO leads to mGroups not being cleared from the system, affecting subsequent mGroup status determination
The above logic leads to: special scenario where mWifiNative.p2pGroupRemove is called and WifiP2pServiceImpl mGroup is not cleared. The result is that the mGroup information does not match the state of the Hal layer, and the upper layer using the WifiP2pManager#requestGroupInfo API can still get the information, which is considered to be a valid GroupInfo.

Suggest to arrange to optimize this issue from system level as well: when mWifiNative.p2pGroupRemove is called, make sure that the mGroup cache information is also cleared at the same time, to avoid the problem of the information and the state not being Match.


Bug:336469699
Change-Id: I0173a74dff7c9f9c7d65b02be465b5a1b8440dd1
Signed-off-by: Peiyong Wang <wangpeiyong@xiaomi.corp-partner.google.com>